backend/requirements.txt
flask
transformers
torch
stable-baselines3
gym
pandas
numpy
requests
python-dotenv